my-chat-ai/requirements.txt
transformers
torch
fastapi
uvicorn
datasets
accelerate
discord.py
requests
python-dotenv